Bez zbędnych plików
