--- Mój_pierwszyNeh/data.txt
@@ -1,5 +1,0 @@
-4 3
-1 3 8
-9 3 5
-7 8 6
-4 8 7
