--- Norbert_lab1/Nowy dokument sformatowany.rtf
@@ -1,1 +1,0 @@
-{\rtf1}
